+++ flutter/example/lib/main.dart
@@ -26,6 +26,7 @@ Future<void> main() async {
       options.reportPackages = false;
       options.addInAppInclude('sentry_flutter_example');
       options.considerInAppFramesByDefault = false;
+      options.attachThreads = true;
     },
     // Init your App.
     appRunner: () => runApp(