+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Feat: Attach Isolate name to thread context (#847)
 * Fix: Fix `SentryAssetBundle` on Flutter >= 3.1 (#877)
 * Feat: Add Android thread to platform stacktraces (#853)
 * Fix: Rename auto initialize property (#857)